small security changes

+++ be/livepark/livepark/src/main/java/ro/license/livepark/controller/parking/ParkingController.java
@@ -6,6 +6,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ro.license.livepark.dto.parking.ParkingDTO;
 import ro.license.livepark.dto.parking.ParkingInfoDTO;
+import ro.license.livepark.entities.parking.Parking;
+import ro.license.livepark.entities.user.User;
 import ro.license.livepark.entities.user.UserRole;
 import ro.license.livepark.service.parking.ParkingService;
 import ro.license.livepark.service.user.UserService;
@@ -20,20 +22,27 @@ public class ParkingController {
     @Resource
     private ParkingService parkingService;
 
+    @Resource
+    private UserService userService;
+
     @GetMapping
     public List<ParkingInfoDTO> getAllParkingsInfo() {
         return parkingService.getAllParkingsInfo();
     }
 
     @PostMapping
-    public ResponseEntity<?> addParking(@RequestBody ParkingDTO parkingDTO, @RequestParam("userId") Long adminId) {
-        int id = parkingService.addParking(parkingDTO, adminId);
+    public ResponseEntity<?> addParking(@RequestBody ParkingDTO parkingDTO) {
+        User authenticatedUser = userService.getAuthenticatedUser();
+        if (authenticatedUser.getRole() == UserRole.USER)
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        int id = parkingService.addParking(parkingDTO, authenticatedUser.getUserId());
         return new ResponseEntity<>(id, HttpStatus.CREATED);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<?> getParking(@PathVariable("id") int id) {
-       ParkingDTO p = parkingService.getParking(id);
+       ParkingDTO p = parkingService.getParkingDTO(id);
        if (p == null)
            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
        return new ResponseEntity<>(p, HttpStatus.OK);
@@ -41,6 +50,8 @@ public class ParkingController {
 
     @PutMapping("/{id}")
     public ResponseEntity<?> modifyParking(@PathVariable("id") int id, @RequestBody ParkingDTO parkingDTO) {
+        if (isUnauthorized(id))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         boolean ok = parkingService.modifyParking(id, parkingDTO);
         if (ok)
             return new ResponseEntity<>(null, HttpStatus.OK);
@@ -49,10 +60,20 @@ public class ParkingController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteParking(@PathVariable("id") int id) {
+        if (isUnauthorized(id))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         boolean ok = parkingService.deleteParking(id);
         if (ok)
             return new ResponseEntity<>(null, HttpStatus.OK);
         return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+    }
+
+    private boolean isUnauthorized(@PathVariable("id") int id) {
+        User authenticatedUser = userService.getAuthenticatedUser();
+        Parking p = parkingService.getParking(id);
+        boolean isParkingAdmin = (p != null && p.getAdminId() == authenticatedUser.getUserId());
+        return authenticatedUser.getRole() == UserRole.USER ||
+                (authenticatedUser.getRole() == UserRole.ADMIN && !isParkingAdmin);
     }
 
 }
